@@ -15,7 +15,7 @@ topics:
   - table
 
 environment:
-  sdk: ">=2.12.0 <4.0.0"
+  sdk: ">=2.15.0 <4.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
