@@ -64,6 +64,7 @@ class _HomePageState extends State<HomePage> {
     }
     return DaviModel(
         rows: characters,
+
         columns: _buildColumns(),
         multiSortEnabled: _multipleSort);
   }
@@ -75,7 +76,9 @@ class _HomePageState extends State<HomePage> {
         leading: const Icon(Icons.person, size: 16),
         name: 'Name',
         width: 100,
-        cellValue: (row, rowIndex) => row.name));
+        rowSpan: (c,index)=>index==_model!.rowsLength-2?2:1,
+        cellValue: (row, rowIndex) => row.name)
+    );
     list.add(DaviColumn(
         pinStatus: _leftPinned ? PinStatus.left : PinStatus.none,
         name: 'Gender',
@@ -86,7 +89,9 @@ class _HomePageState extends State<HomePage> {
             : CellIcon(Icons.female, color: Colors.pink[600]!)));
     list.add(
         DaviColumn(name: 'Race', width: 100,
-    cellValue: (row, rowIndex)=> row.race));
+    //cellValue: (row, rowIndex)=> row.race)
+            cellWidget: (d,qd,t)=>TextButton(onPressed: ()=>print('a'), child: Text('a')))
+    );
     list.add(
         DaviColumn(name: 'Class', width: 110,
     cellValue: (row,rowIndex)=>row.cls));
@@ -139,6 +144,7 @@ class _HomePageState extends State<HomePage> {
 
     Widget table = DaviTheme(
         child: Davi<Character>(_model!,
+            //collisionBehavior: CellCollisionBehavior.throwException,
             columnWidthBehavior: _columnsFit?ColumnWidthBehavior.fit:ColumnWidthBehavior.scrollable,
             rowColor: _rowColor
                 ? (data, rowIndex, hovered) =>
