@@ -4,15 +4,11 @@ import 'package:meta/meta.dart';
 /// Parent data for [CellsLayoutRenderBox] class.
 @internal
 class CellsLayoutParentData extends ContainerBoxParentData<RenderBox> {
-  int? rowIndex;
-  int? columnIndex;
-  int? rowSpan;
-  int? columnSpan;
+  int? cellIndex;
 
   bool get isCell =>
-      rowIndex != null &&
-      rowIndex! >= 0 &&
-      columnIndex != null &&
-      columnIndex! >= 0;
+      cellIndex != null &&
+          cellIndex! >= 0;
+
   bool get isTrailing => !isCell;
 }
