@@ -2,7 +2,6 @@ import 'package:davi/src/column.dart';
 import 'package:davi/src/internal/columns_layout.dart';
 import 'package:davi/src/internal/columns_layout_child.dart';
 import 'package:davi/src/internal/new/davi_context.dart';
-import 'package:davi/src/internal/scroll_offsets.dart';
 import 'package:davi/src/internal/table_layout_settings.dart';
 import 'package:davi/src/theme/theme.dart';
 import 'package:davi/src/theme/theme_data.dart';
@@ -12,13 +11,11 @@ class SummaryWidget<DATA> extends StatelessWidget {
   const SummaryWidget(
       {Key? key,
       required this.daviContext,
-      required this.layoutSettings,
-      required this.horizontalScrollOffsets})
+      required this.layoutSettings})
       : super(key: key);
 
   final DaviContext<DATA> daviContext;
   final TableLayoutSettings layoutSettings;
-  final HorizontalScrollOffsets horizontalScrollOffsets;
 
   @override
   Widget build(BuildContext context) {
@@ -51,7 +48,7 @@ class SummaryWidget<DATA> extends StatelessWidget {
 
     Widget summary = ColumnsLayout(
         layoutSettings: layoutSettings,
-        horizontalScrollOffsets: horizontalScrollOffsets,
+        scrollControllers: daviContext.scrollControllers,
         columnDividerThickness: theme.columnDividerThickness,
         columnDividerColor: theme.summary.columnDividerColor,
         children: children);
