@@ -1,21 +1,16 @@
 import 'package:davi/davi.dart';
 import 'package:davi/src/column.dart';
-import 'package:davi/src/column_width_behavior.dart';
 import 'package:davi/src/internal/column_metrics.dart';
 import 'package:davi/src/internal/header_widget.dart';
 import 'package:davi/src/internal/layout_child_id.dart';
 import 'package:davi/src/internal/new/davi_context.dart';
 import 'package:davi/src/internal/new/summary_widget.dart';
 import 'package:davi/src/internal/new/table_content.dart';
-import 'package:davi/src/internal/scroll_controllers.dart';
-import 'package:davi/src/internal/scroll_offsets.dart';
 import 'package:davi/src/internal/table_edge.dart';
 import 'package:davi/src/internal/table_layout.dart';
 import 'package:davi/src/internal/table_layout_child.dart';
 import 'package:davi/src/internal/table_layout_settings.dart';
 import 'package:davi/src/internal/table_scrollbar.dart';
-import 'package:davi/src/theme/theme.dart';
-import 'package:davi/src/theme/theme_data.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
@@ -24,12 +19,10 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
   const TableLayoutBuilder(
       {Key? key,
       required this.daviContext,
-      required this.scrollControllers,
       required this.onDragScroll})
       : super(key: key);
 
   final DaviContext<DATA> daviContext;
-  final ScrollControllers scrollControllers;
   final OnDragScroll onDragScroll;
 
   @override
@@ -39,9 +32,6 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
 
   Widget _builder(BuildContext context, BoxConstraints constraints) {
     final DaviThemeData theme = DaviTheme.of(context);
-
-    final HorizontalScrollOffsets horizontalScrollOffsets =
-        HorizontalScrollOffsets(scrollControllers);
 
     TableLayoutSettings layoutSettings = TableLayoutSettings(
         constraints: constraints,
@@ -74,7 +64,7 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
           child: TableScrollbar(
               axis: Axis.vertical,
               contentSize: layoutSettings.contentHeight,
-              scrollController: scrollControllers.vertical,
+              scrollController: daviContext.scrollControllers.vertical,
               color: theme.scrollbar.verticalColor,
               borderColor: theme.scrollbar.verticalBorderColor,
               onDragScroll: onDragScroll)));
@@ -87,8 +77,7 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
               daviContext: daviContext,
               layoutSettings: layoutSettings,
               resizable: daviContext.columnWidthBehavior ==
-                  ColumnWidthBehavior.scrollable,
-              horizontalScrollOffsets: horizontalScrollOffsets)));
+                  ColumnWidthBehavior.scrollable)));
       if (layoutSettings.hasVerticalScrollbar) {
         children.add(TableLayoutChild(
             id: LayoutChildId.headerEdge,
@@ -99,7 +88,7 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
     if (layoutSettings.hasHorizontalScrollbar) {
       Widget leftPinnedHorizontalScrollbar = TableScrollbar(
           axis: Axis.horizontal,
-          scrollController: scrollControllers.leftPinnedHorizontal,
+          scrollController: daviContext.scrollControllers.leftPinnedHorizontal,
           color: theme.scrollbar.pinnedHorizontalColor,
           borderColor: theme.scrollbar.pinnedHorizontalBorderColor,
           contentSize: layoutSettings.leftPinnedContentWidth,
@@ -111,7 +100,7 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
           id: LayoutChildId.unpinnedHorizontalScrollbar,
           child: TableScrollbar(
               axis: Axis.horizontal,
-              scrollController: scrollControllers.unpinnedHorizontal,
+              scrollController: daviContext.scrollControllers.unpinnedHorizontal,
               color: theme.scrollbar.unpinnedHorizontalColor,
               borderColor: theme.scrollbar.unpinnedHorizontalBorderColor,
               contentSize: layoutSettings.unpinnedContentWidth,
@@ -125,19 +114,24 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
 
     children.add(TableLayoutChild(
         id: LayoutChildId.cells,
-        child: TableContent(
+        child:
+        LayoutBuilder(builder:(context, constraints) {
+          DaviThemeData theme = DaviTheme.of(context);
+        return TableContent(
             daviContext: daviContext,
             layoutSettings: layoutSettings,
-            horizontalScrollOffsets: horizontalScrollOffsets,
-            verticalScrollController: scrollControllers.vertical)));
+        rowFillHeight: theme.row.fillHeight,
+        maxWidth: constraints.maxWidth,
+        maxHeight: constraints.maxHeight);
+        }
+    )));
 
     if (daviContext.model.hasSummary) {
       children.add(TableLayoutChild(
           id: LayoutChildId.summary,
           child: SummaryWidget(
               daviContext: daviContext,
-              layoutSettings: layoutSettings,
-              horizontalScrollOffsets: horizontalScrollOffsets)));
+              layoutSettings: layoutSettings)));
       if (layoutSettings.hasVerticalScrollbar) {
         children.add(TableLayoutChild(
             id: LayoutChildId.summaryEdge,
@@ -148,7 +142,6 @@ class TableLayoutBuilder<DATA> extends StatelessWidget {
     return TableLayout<DATA>(
         layoutSettings: layoutSettings,
         theme: theme,
-        horizontalScrollOffsets: horizontalScrollOffsets,
         children: children);
   }
 }
